+++ core/src/index/knowhere/knowhere/index/vector_index/helpers/IndexParameter.h
@@ -52,6 +52,8 @@ constexpr const char* L2 = "L2";
 constexpr const char* HAMMING = "HAMMING";
 constexpr const char* JACCARD = "JACCARD";
 constexpr const char* TANIMOTO = "TANIMOTO";
+constexpr const char* SUBSTRUCTURE = "SUBSTRUCTURE";
+constexpr const char* SUPERSTRUCTURE = "SUPERSTRUCTURE";
 }  // namespace Metric
 
 extern faiss::MetricType