+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Please mark all change in change log and use the issue from GitHub
 -   \#1648 The cache cannot be used all when the type is binary
 
 ## Feature
+-   \#1603 BinaryFlat add 2 Metric: Substructure and Superstructure
 
 ## Improvement
 -   \#1537 Optimize raw vector and uids read/write